Language.Haskell.TH.Specialize

Main Interface

expand_and_specialize :: Name -> Name -> Q [Dec]

Expand all the type syn's and create specialize types for any polymorphic types. All of the new specialized declarations are returned, along with the original dec with subbed types and a new name. The first Name is the name of the Dec to create specialize instances for. The second Name, is the new name for the Dec. use mk_new_dec_name for the Dec renaming and id_constr_renamer for the constructor renaming.

expand_and_specialize' :: DecRenamer -> ConstrRenamer -> Name -> Name -> Q [Dec]

Expand all the type syn's and create specialize types for any polymorphic types. All of the new specialized declarations are returned, along with the original dec with subbed types and a new name. The first Name is the name of the Dec to create specialize instances for. The second Name, is the new name for the Dec. The DecRenamer and ConstrRenamer are used to rename Dec's and Con's respectively.

Helper Types

newtype ConstructorName

Constructors

ConstructorName

Fields

runConstructorName :: Name

newtype TypeName

Constructors

TypeName

Fields

runTypeName :: Name

Renamer Interface

type DecRenamer = [Type] -> TypeName -> Result TypeName

type ConstrRenamer = [Type] -> Con -> Con

Stock Renamers

mk_new_dec_name :: [Type] -> TypeName -> Result TypeName

Default Dec renamer

id_constr_renamer :: [Type] -> Con -> Con

Default Con renamer

Utils ... Pretend these are not here

sub_dec_and_rename :: (Monad m, Functor m, MonadError String m) => ConstrRenamer -> Dec -> [Type] -> m Dec

create_dec_from_type :: (Functor m, Monad m, MonadState [Dec] m, MonadError String m) => DecRenamer -> ConstrRenamer -> Type -> m Type

find_con :: (Monad m, MonadError String m) => ConstructorName -> Dec -> m Con

get_con_vars :: Con -> [Type]

get_ty_vars :: Dec -> [TyVarBndr]

concat_type_names :: [Type] -> String

rename_dec :: (Monad m, MonadError String m) => [Type] -> Dec -> m Dec

run_state' :: ErrorStateT e s m a -> s -> m (Either e a, s)

type Result a = Either String a